main scene, basic ui, tool type
This is before going through the "Resource Gathering RPG in Godot 4" tutorial by Chris' Tutorials.
Created a main scene to use as my base map for testing.
Created UI file format and update hotbar buttons. hotbar buttons do not currently incorporate the tools, but has the ability to if the tools are in scope of the hotbar, which did not seem like the correct way to go about it.
Made the Tool type, which has basic stats such as quality, strength, and range. This type should include all common functionality between different tools.
Made the pickaxe and woodcutting axe both of type Tool.