Interact Menu QoL, StatusEffect creation
made the Interact Menu close when you leave the radius or when you press the interact button a second time.
Created StatusEffect class that is a template for the possible Status Effects the player can have. Player requires to have the StatusEffect in the known list of status effects to be able to have an effect, as each status effect will have a unique function for the player and will be written separately