batch changes
forgot to update for at least a week.
Added inventory. Grabbed items from context menu on world objects get added to the inventory.
Added StatusEffects. When consumed, items will trigger status effects on the player that alter the characters "stats".
Added the panel for AlchemyActivity, which is where the alchemy minigame will occur for now. Items from inventory can be dragged using left click and dropped onto the cauldron to transfer the item from the inventory to the cauldron.
Added associated assets.
Added concept art and notes.
ContextButtons in the ContextBox for world objects can now be generated at runtime to allow more flexibility in the displayed contents of the ContextBox.