Mortar and Pestle, Object duplication, Status Effect Icons
Added the Mortar and Pestle to the alchemy minigame.
Adjusted how signals are sent from objects and how they handle the quantities of their stored items so that they do not share the same Resource.
Added icons for Status Effects that occur to the player as well as a progress bar for the duration of the effect