sliding door puzzle
added functionality to make a door slide open and closed when sent the proper signals.
Made a pressure plate that sends signals to a door to open and close it.
Created red berry bush items that give red berries that can remove statuses and are an ingredient in health potions