finished hotbar, recipe .tres, heirarchy update
Hotbar now consumes the item for the associated button pressed. If there are no more items in the inventory when the hotbar button is pressed, the item is removed from the hotbar. (Note, this might not handle when the item is removed from the inventory from outside the hotbar).
Set up CauldronRecipe and MortarPestleRecipe classes and resource generation. This should allow recipes to be added more easily.
Added more folders for organization.