Options Menu, Global Modes, New Recipes,  Misc.
Added an options menu with the ability to return to the game, open a settings menu (currently not functional), and exit the application.
Added a global "mode" to keep track of which UI content is showing (and perhaps for other uses in the future).
Added more items for grinding up the berries and flowers as well as associated recipes. Added recipe for cleanse potion.
Automated the Recipe collection for each alchemy tool so that as long as the recipe is in the correct folder, it will be added to the list of recipes for the associated tool.

Other miscellaneous changes.