Added M&P Minigame, Procedures, Updated Recipe
Added Procedure type which contains a type ProcedureInputActions to keep track of the input at each step of the procedure.
Mortar and Pestle minigame only works with the green herb recipe as no other procedures have been configured.
Made Items "local to scene", but no updates to Items besides that.