Minigame Button Icons, Menu Bar,  Canvas Reparent
Added icons for buttons in the minigames and recipe book.
Added a menu bar to allow users to press on the menu buttons to access the inventory, recipe book, and options menu.
Moved the Canvas UI scenes that were children of the Player and made them children of the world instead. Code mostly unchanged except for re-linking a signal.
Added tooltips to the selected tool in the tool wheel as well as the menu bar buttons with their hotkeys shown.