UI fixes, minigame logic, new items
Adjusted UI to clear clutter and to allow the recipe list to be shown alongside the crafting and inventory.
Adjusted minigame logic to check for the item IDs instead of the slot used, preventing errors when objects are included in the minigame but not used.
Began implementation of new item and recipe set which should have a consistent logic for procedures. This should allow for a sensible way to give the player hints based on the rules of how the procedures are designed.
Misc bug fixes