Interactable_plant reformat
Changed the file and data structure of interactable plants to allow streamlined editing.
Added flower stem to the list of items that can be gathered from the yellow flower. There is currently no indication to the player as to why the plant does not disappear if they are only grabbing the flower.