Stat, potion inventory, recipe updates
Added "mass" to Player to be used to determine potential future interactions with the environment and the player.
Updated the Potion of Growth to work with the potion item changes.
Made the inventory items an exported variable and any items in the inventory when the game loads will now be added to the inventory.
Fixed merger recipes not being compatible with the recipe list.
Added display for ingredients to the recipe list when the recipe is not using the cauldron. No item counts since only a single amount of each ingredient is needed.